[refactor]: 회원가입 컴포넌트 사용 위치에 따라 common, signup 으로 이동 #61
