+++ app/src/main/java/com/kuit/ourmenu/ui/common/LoginTextField.kt
@@ -1,4 +1,4 @@
-package com.kuit.ourmenu.ui.onboarding.component
+package com.kuit.ourmenu.ui.common
 
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.PaddingValues
@@ -19,7 +19,6 @@ import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.kuit.ourmenu.ui.common.CustomTextField
 import com.kuit.ourmenu.ui.theme.Neutral100
 import com.kuit.ourmenu.ui.theme.Neutral300
 import com.kuit.ourmenu.ui.theme.Neutral500
